xml中某一个节点本应该是一个数组，但是如果数组中只有一个元素，就只会被解析成一个对象，所以需要把该节点指定成数组
